requirements.txt
# Flask Framework
Flask==1.1.2
Flask-Mail==0.9.0
Flask-Testing-0.8.1
APScheduler==3.8.0
Werkzeug==1.0.1
itsdangerous==1.1.0

# PostGres database
psycopg2-binary==2.8.6
requests~=2.18.4
bcrypt~=3.2.0